fix "arrays used as indices must be of integer (or boolean) type

+++ ex7_K_means_Clustering_and_Principal_Component_Analysis/findClosestCentroids.py
@@ -14,7 +14,7 @@ def findClosestCentroids(X, centroids):
     K, _ = np.shape(centroids)
 
     # You need to return the following variables correctly. idx = zeros(size(X, 1), 1);
-    idx = np.zeros((np.shape(X)[0], 1))
+    idx = np.zeros((np.shape(X)[0], 1), dtype=np.int8)
 
     # ====================== YOUR CODE HERE ======================
     # Instructions: Go over every example, find its closest centroid, and store
